make refactor to rename files and reorganize folder structure
