+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/mediadetails/details/MediaInteractionListener.kt
@@ -1,9 +1,9 @@
-package com.berlin.aflami.viewmodel.mediadetails
+package com.berlin.aflami.viewmodel.mediadetails.details
 
 import com.berlin.aflami.viewmodel.reusableinteractionlistener.addtofavourite.AddToFavouriteInteractionListener
 import com.berlin.aflami.viewmodel.reusableinteractionlistener.createnewlist.CreateNewListInteractionListener
 import com.berlin.aflami.viewmodel.reusableinteractionlistener.rate.RateInteractionListener
-import com.berlin.aflami.viewmodel.mediadetails.uistate.MediaType
+import com.berlin.aflami.viewmodel.shareduistate.MediaType
 
 interface MediaInteractionListener :
     RateInteractionListener,
@@ -19,7 +19,7 @@ interface MediaInteractionListener :
 
 }
 
-interface ExtraMediaContentInteractionListener : TVShowSeasonsInteractionListener  {
+interface ExtraMediaContentInteractionListener : TVShowSeasonsInteractionListener {
     fun onShowMoreMediaLikeThisClicked(mediaId: Long, mediaType: MediaType)
     fun onShowReviewsClicked(mediaId: Long, mediaType: MediaType)
     fun onShowMediaGalleryClicked(id:Long,mediaType: MediaType)
