+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/shareduistate/MediaUiState.kt
@@ -1,0 +1,13 @@
+package com.berlin.aflami.viewmodel.shareduistate
+
+enum class MediaType { MOVIE, TV_SHOW }
+
+data class MediaUiState(
+    val id: Long = 0L,
+    val title: String = "",
+    val rating: String = "",
+    val releaseYear: String = "",
+    val genre: List<Int> = emptyList(),
+    val poster: String = "",
+    val mediaType:MediaType=MediaType.MOVIE
+)
